finish updating PO schema and API according to new specs

+++ Controller/PurchaseOrderController.js
@@ -1,10 +1,26 @@
 const purchaseOrderDAO = require("../DAO/PurchaseOrderDAO");
+const userController = require("../Controller/UserController");
 const net = require("../Util/Net");
+const counter = require("../Model/Counter");
 
-//todo: implement permissions check using session (see UserController.js)
+const allowed_fields = ["owner", "description", "parts", "status", "subteam", "deadline", "priority", "comment", "total_price"];
+exports.allowed_fields = allowed_fields;
+
+const required_fields = ["owner", "description", "parts", "subteam", "deadline", "priority", "total_price"];
+exports.required_fields = required_fields;
 
 exports.getAll = (req, res) => {
     console.log("API GET request called for all purchase orders");
+
+    if (!req.user || !req.isAuthenticated()) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request; please login"));
+        return;
+    }
+
+    if (userController.permission_levels[req.user.role] < userController.permission_levels["member"]) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request"));
+        return;
+    }
 
     purchaseOrderDAO.getAllPurchaseOrders().then(function (pos) {
         res.json(net.getSuccessResponse(null, pos));
@@ -14,8 +30,18 @@ exports.getAll = (req, res) => {
     });
 };
 
-exports.getPO = (req, res) => {
+exports.get = (req, res) => {
     console.log(`API GET request called for purchase order ${req.params.num}`);
+
+    if (!req.user || !req.isAuthenticated()) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request; please login"));
+        return;
+    }
+
+    if (userController.permission_levels[req.user.role] < userController.permission_levels["member"]) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request"));
+        return;
+    }
 
     purchaseOrderDAO.getPO(req.params.num).then(function (po) {
         res.json(net.getSuccessResponse(null, po));
@@ -25,46 +51,118 @@ exports.getPO = (req, res) => {
     });
 };
 
-//todo rehaul
 exports.update = (req, res) => {
     console.log(`API PUT request called for purchase order ${req.params.num}`);
 
-    const params = req.body;
+    if (!req.user || !req.isAuthenticated()) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request; please login"));
+        return;
+    }
+
+    //todo: keep array of po_nums in user obj to quickly check if this PO belongs to requesting user
+    // if (req.params.email !== req.user.email && userController.permission_levels[req.user.role] < userController.permission_levels["manager"]) {
+    //     res.status(401).json(net.getErrorResponse("you are not authorized to make this request; must be your PO or must be at least manager"));
+    //     return;
+    // }
+
+    if (userController.permission_levels[req.user.role] < userController.permission_levels["manager"]) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request; must be at least manager"));
+        return;
+    }
 
     //assume parameters have been sanitized on client side
+    const params = req.body;
+    const keys = Object.keys(params);
 
-    //todo: add subteam
-    if (Object.keys(params).length === 4) {
-        purchaseOrderDAO.updatePO(req.params.num, params["po_number"], params["owner"], params["description"], params["file_location"]).then(function (updatedPO) {
-            console.log("PO " + updatedPO.num + " Updated!", updatedPO);
-            res.json(net.getSuccessResponse("updated", updatedPO));
-        }).catch(function (err) {
-            console.log("failed to update record");
-            if (err.name === "ValidationError") {
-                console.error("Error Validating!", err);
-                res.status(422).json(net.getErrorResponse(err));
-            } else {
-                console.error(err);
-                res.status(500).json(net.getErrorResponse("failed to update record"));
-            }
-        });
-    } else {
-        res.status(404).json(net.getErrorResponse("Insufficient parameters provided"));
+    if (keys.length === 0) {
+        res.status(404).json(net.getErrorResponse("update request must include at least on parameter"));
+        return;
     }
+
+    for (let i = 0; i < keys.length; ++i) {
+        if (!allowed_fields.includes(keys[i])) {
+            res.status(404).json(net.getErrorResponse(`cannot update field '${keys[i]}' or does not exist`));
+            return;
+        }
+    }
+
+    //validate parts
+    if (params.parts) {
+        //todo
+    }
+
+    purchaseOrderDAO.updatePO(req.params.num, params).then(function (updatedPO) {
+        console.log("PO " + updatedPO.num + " Updated!", updatedPO);
+        res.json(net.getSuccessResponse("updated", updatedPO));
+    }).catch(function (err) {
+        console.log("failed to update record");
+        if (err.name === "ValidationError") {
+            console.error("Error Validating!", err);
+            res.status(422).json(net.getErrorResponse(err));
+        } else {
+            console.error(err);
+            res.status(500).json(net.getErrorResponse("failed to update record"));
+        }
+    });
 };
 
 exports.create = (req, res) => {
     console.log(`API POST request called for "create PO"`);
 
-    const params = req.body;
+    if (!req.user || !req.isAuthenticated()) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request; please login"));
+        return;
+    }
+
+    if (userController.permission_levels[req.user.role] < userController.permission_levels["member"]) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request"));
+        return;
+    }
 
     //assume parameters have been sanitized on client side
+    const params = req.body;
+    const keys = Object.keys(params);
 
-    //po_number will be generated server side
-    //subteam will be gotten from user session
+    //if owner not set, get from user session
+    if (!params.owner) {
+        params.owner = req.user.email;
+    }
 
-    if (Object.keys(params).length === 4) {
-        purchaseOrderDAO.createPurchaseOrder(params["po_number"], params["owner"], params["description"], params["file_location"]).then(function(newPO) {
+    //if subteam not set, get from user session
+    if (!params.subteam) {
+        params.subteam = req.user.subteam;
+    }
+
+    if (keys.length < required_fields.length) {
+        res.status(404).json(net.getErrorResponse("Insufficient parameters provided"));
+        return;
+    }
+
+    //check required fields
+    for (let i = 0; i < required_fields.length; ++i) {
+        if (!keys.includes(required_fields[i])) {
+            res.status(404).json(net.getErrorResponse(`'${required_fields[i]}' field is required`));
+            return;
+        }
+    }
+
+    //check other fields allowed
+    for (let i = 0; i < keys.length; ++i) {
+        if (!allowed_fields.includes(keys[i])) {
+            res.status(404).json(net.getErrorResponse(`cannot set field '${keys[i]}' or does not exist`));
+            return;
+        }
+    }
+
+    //validate parts
+    if (params.parts) {
+        //todo
+    }
+
+    counter.getNextSequenceValue("po_counter").then(function (updatedCounter) {
+        params.po_number = updatedCounter.sequence_value;
+
+        purchaseOrderDAO.createPurchaseOrder(params).then(function(newPO) {
             console.log("New PO Created!", newPO);
             res.json(net.getSuccessResponse("new po created", newPO));
         }).catch(function(err) {
@@ -76,13 +174,31 @@ exports.create = (req, res) => {
                 res.status(500).json(net.getErrorResponse(err));
             }
         });
-    } else {
-        res.status(404).json(net.getErrorResponse("Insufficient parameters provided"));
-    }
+    }).catch(function (err) {
+        console.log("could not get next po number from counter: ", err);
+        console.error("could not create PO");
+        res.status(500).json(net.getErrorResponse(err));
+    });
 };
 
 exports.delete = (req, res) => {
     console.log(`API DELETE request called for purchase order ${req.params.num}`);
+
+    if (!req.user || !req.isAuthenticated()) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request; please login"));
+        return;
+    }
+
+    //todo: keep array of po_nums in user obj to quickly check if this PO belongs to requesting user
+    // if (req.params.email !== req.user.email && userController.permission_levels[req.user.role] < userController.permission_levels["manager"]) {
+    //     res.status(401).json(net.getErrorResponse("you are not authorized to make this request; must be your PO or must be at least manager"));
+    //     return;
+    // }
+
+    if (userController.permission_levels[req.user.role] < userController.permission_levels["manager"]) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request; must be at least manager"));
+        return;
+    }
 
     purchaseOrderDAO.deletePO(req.params.num).then(function (result) {
         if (result.deletedCount === 0) {
@@ -90,7 +206,7 @@ exports.delete = (req, res) => {
             res.status(404).json(net.getErrorResponse("could not find record to remove for po number: " + req.params.num));
         } else if (result.deletedCount === 1) {
             //success
-            res.json(net.getSuccessResponse("successfully removed record", req.params.num))
+            res.json(net.getSuccessResponse("successfully removed record", {po_number: req.params.num}))
         } else {
             //critical error
             res.status(500).json(net.getErrorResponse("critical server error"));
@@ -104,6 +220,16 @@ exports.delete = (req, res) => {
 exports.getBySubteam = (req, res) => {
     console.log(`API GET request called for purchase order from subteam ${req.params.subteam}`);
 
+    if (!req.user || !req.isAuthenticated()) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request; please login"));
+        return;
+    }
+
+    if (userController.permission_levels[req.user.role] < userController.permission_levels["member"]) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request"));
+        return;
+    }
+
     purchaseOrderDAO.getBySubteam(req.params.subteam).then(function (pos) {
         res.json(net.getSuccessResponse(null, pos));
     }).catch(function (err) {
@@ -115,10 +241,24 @@ exports.getBySubteam = (req, res) => {
 exports.getByUser = (req, res) => {
     console.log(`API GET request called for purchase orders from user ${req.params.email}`);
 
+    if (!req.user || !req.isAuthenticated()) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request; please login"));
+        return;
+    }
+
+    if (userController.permission_levels[req.user.role] < userController.permission_levels["member"]) {
+        res.status(401).json(net.getErrorResponse("you are not authorized to make this request"));
+        return;
+    }
+
     purchaseOrderDAO.getByUser(req.params.email).then(function (pos) {
         res.json(net.getSuccessResponse(null, pos));
     }).catch(function (err) {
         console.log("error getting purchase orders: ", err);
         res.status(500).json(net.getErrorResponse("error retrieving records from database"));
     });
+};
+
+exports.updateStatus = (req, res) => {
+  //todo
 };
